Created get_dev_train_split.py
Added python script that takes a directory with mrp corpus files (say amr graphs) and produces two files (in two different folders), namely the dev and the training corpus, both in mrp format. Ids of the graphs were changed to also contain the filename. Assumes that each line of the input files is one graph and starts with the id of that graph!