adding ignore aware outside estimator, also in comment: use  Astar from master instead
the default estimator is the static one, but ignore-aware should be better.
On the cogs branch Astar is still in experimental. On the master branch it was moved out of experimental and refactored afaik. If we want use the code form the master branch we woud need a separate, second jar file...